requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
humanfriendly==10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
protobuf==5.28.3
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rich==13.9.3
rpds-py==0.20.1
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2024.8.30
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.0
wrapt==1.16.0
zipp==3.20.2
